requirements.txt
tokenizers
transformers
tqdm 
indic-nlp-library
torch 
torchvision 
torchaudio
accelerate
numpy
pandas
datasets
sacremoses
sentencepiece
